requirements.txt
dotenv
requests
pandas
google-cloud-bigquery
faker
pendulum
apache-airflow-providers-google
openpyxl